Bring back ImageJMain class, use in ImageJMainTest class
